Rating front finished

+++ AccommodationBackend/rating_service/domain/repository/accommodation_rating_repository.go
@@ -51,7 +51,7 @@ func (a AccommodationMongoDBStore) GetAverageScoreForAccommodation(accommodation
 		return 0, err
 	}
 	if len(result) > 0 {
-		return result[0]["averageScore"].(float32), nil
+		return float32(result[0]["averageScore"].(float64)), nil
 	}
 	return 0.0, nil
 }
